Refactor: Split into several classes, use activity classes
This is a large refactoring, mostly for usage of the new activity
classes. Since a DSL around tasks isn't flexible enough for the
new activity classes this also revamps the whole DSL.
It also adds a matchobject class for rule conditionals which
should allow to trace problems with the dsl.